Update setting-default.md
Downloaded the extension and tried the chromium based browsers instructions as well. The "make default" option was disabled. 
Tried testing with a search and it seemed to work. Might need to add the popup that asked if I wanted to keep Kagi as default or change back to chrome. Idk if that was just because of the extension or not but decided to make a couple small changes I noticed through the process anyways.